as of perl 5.8.3 in-memory-filehandles have problems with STD streams and threads. A sequence:
  % t/TEST -v t/modperl/request_rec_perlio_api.t \
  t/modperl/io_nested_with_closed_stds.t

was failing because of the warning in the latter test:

Filehandle STDOUT reopened as STDIN only for input at
/home/stas/apache.org/modperl-2.0/t/response/TestModperl/io_nested_with_closed_stds.pm
line 54.

This is because PerlIO_stdin() wasn't getting restored after the
in-memory-filehandle was used and had it fileno stuck at that
in-memory-filehandle that has gone out of scope.

Solution: use a plain file to reopen STDIN

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/perl/modperl/trunk@70653 13f79535-47bb-0310-9956-ffa450edef68