Fix a crash when running a sub-request from within a filter where mod_perl was not the content handler.
Running modperl_wbucket_flush(rcfg->wbucket) before the subrequest
is run ensures we've flushed our own content handler if it has
any data in the pipeline. Problem was rcfg->wbucket will not be
set if we are not the content handler.

Solution is to avoid the flush in that particular case.



git-svn-id: https://svn.apache.org/repos/asf/perl/modperl/trunk@607687 13f79535-47bb-0310-9956-ffa450edef68