Add a Changes entry for the biggest change in this version, which is as-yet unmentioned in the Change file!
If my reading of the SVN history is correct, this was started by Torsten, finished for *nix by Jan, largely fixed up for Windows by me, and with a few contributions from gozer. My apologies if I've missed anyone. If so then speak up during the RC phase to get your name in lights and receive fame, fortune and probably more spam.

git-svn-id: https://svn.apache.org/repos/asf/perl/modperl/trunk@1679260 13f79535-47bb-0310-9956-ffa450edef68