Automatic compiler flags for gcc 5 and clang
To make compilation easier on gcc5/clang hosts, this patch automatically selects the appropriate c89 option, when modperl is being built with either gcc 5 or clang.

Tested by the author on Ubuntu 15.10 (with gcc 5.2.1), Fedora 23 (with gcc 5.3.1) and FreeBSD 10.2 (with clang).

Thanks to Klaus S. Madsen <ksm@jobindex.dk> for the patch.

git-svn-id: https://svn.apache.org/repos/asf/perl/modperl/trunk@1733566 13f79535-47bb-0310-9956-ffa450edef68